backend/requirements.txt
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
scipy==1.11.4
openai==2.6.1
python-multipart==0.0.6
asyncpg==0.29.0
sentence-transformers==2.2.2
pgvector==0.2.4
supabase==2.11.0
requests==2.31.0
python-dotenv==1.0.1
ijson==3.4.0
neo4j>=5.8.0,<6
passlib[bcrypt]==1.7.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
jsonschema==4.19.0
pytest-asyncio==0.21.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-jaeger==1.21.0
